asdsa

+++ python/test.py
@@ -87,7 +87,7 @@ res=myNet(test)
 
 print(myTools.myTestCostFunction(res,masks[splitPoint+1:masks.shape[0], :, :, :]))
 
-print(sklearn.metrics.log_loss(masks[splitPoint+1:masks.shape[0], :, :, :], res))
+print(sklearn.metrics.log_loss(masks[splitPoint+1:masks.shape[0].astype(numpy.int), :, :, :], res))
 
 
 trainInstance=train[0]
